move checkpointer config and other stuff
